requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile --resolver=backtracking
#
attrs==23.1.0
    # via
    #   cattrs
    #   jsii
aws-cdk-asset-awscli-v1==2.2.198
    # via aws-cdk-lib
aws-cdk-asset-kubectl-v20==2.1.2
    # via aws-cdk-lib
aws-cdk-asset-node-proxy-agent-v5==2.0.165
    # via aws-cdk-lib
aws-cdk-lib==2.85.0
    # via -r requirements.in
boto3==1.26.158
    # via -r requirements.in
botocore==1.29.158
    # via
    #   -r requirements.in
    #   boto3
    #   s3transfer
cattrs==23.1.2
    # via jsii
constructs==10.2.60
    # via aws-cdk-lib
exceptiongroup==1.1.1
    # via cattrs
importlib-resources==5.12.0
    # via jsii
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsii==1.84.0
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
numpy==1.25.0
    # via -r requirements.in
psycopg2-binary==2.9.6
    # via -r requirements.in
publication==0.0.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
pygresql==5.2.4
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   botocore
    #   jsii
s3transfer==0.6.1
    # via boto3
six==1.16.0
    # via python-dateutil
typeguard==2.13.3
    # via
    #   aws-cdk-asset-awscli-v1
    #   aws-cdk-asset-kubectl-v20
    #   aws-cdk-asset-node-proxy-agent-v5
    #   aws-cdk-lib
    #   constructs
    #   jsii
typing-extensions==4.6.3
    # via
    #   cattrs
    #   jsii
urllib3==1.26.16
    # via botocore
